moved irrelated stuff from server to the right locations #93 #94
