+++ src/utils/logger.ts
@@ -1,7 +1,11 @@
 import * as winston from 'winston';
 import * as os from 'os';
 import * as Elasticsearch from 'winston-elasticsearch';
+import * as grpc from 'grpc';
+import apm from 'elastic-apm-node';
 import { confLogger, serviceName } from '../config';
+import { statusToString, validateGrpcError } from './errors/grpc.status';
+import { ApplicationError } from './errors/application.error';
 
 // index pattern for the logger
 const indexTemplateMapping = require('winston-elasticsearch/index-template-mapping.json');
@@ -42,4 +46,44 @@ export enum Severity {
   VERBOSE = 'verbose',
   DEBUG = 'debug',
   SILLY = 'silly',
+}
+
+
+
+  /**
+   * wraps all of the service methods, creating the transaction for the apm and the logger,
+   * and sends them to the elastic server.
+   * @param func - the method called and wrapped.
+   */
+export function wrapper (func: Function) :
+  (call: grpc.ServerUnaryCall<Object>, callback: grpc.requestCallback<Object>) => Promise<void> {
+  return async (call: grpc.ServerUnaryCall<Object>, callback: grpc.requestCallback<Object>) => {
+    try {
+      const traceparent: grpc.MetadataValue[] = call.metadata.get('elastic-apm-traceparent');
+      const transOptions = (traceparent.length > 0) ? { childOf: traceparent[0].toString() } : {};
+      apm.startTransaction(`/file.FileService/${func.name}`, 'request', transOptions);
+      const traceID: string = getCurrTraceId();
+      log(Severity.INFO, func.name, 'request', traceID, call.request);
+
+      const res = await func(call, callback);
+
+      apm.endTransaction(statusToString(grpc.status.OK));
+      log(Severity.INFO, func.name, 'response', traceID, res);
+      callback(null, res);
+    } catch (err) {
+      const validatedErr : ApplicationError = validateGrpcError(err);
+      log(Severity.ERROR, func.name, err.message, getCurrTraceId());
+      apm.endTransaction(validatedErr.name);
+      callback(validatedErr);
+    }
+  };
+}
+
+function getCurrTraceId() : string {
+  try {
+    return apm.currentTransaction.traceparent.split('-')[1];
+  } catch (err) {
+    // Should never get here. The log is set after apm starts.
+    return '';
+  }
 }
